+++ db/migrate/20201203153100_add_active_for_to_tag_type.rb
@@ -1,6 +1,6 @@
 class AddActiveForToTagType < ActiveRecord::Migration[6.0]
   def change
-    add_column :tag_types, :active_for_activity_at, :datetime
-    add_column :tag_types, :active_for_time_range_at, :datetime
+    add_column :tag_types, :active_for_activities_at, :datetime
+    add_column :tag_types, :active_for_time_ranges_at, :datetime
   end
 end